MOTECH-2101: Fixed comment issues
